nostradamIQ-master/demo_app/nostradamIQ/requirements.txt
gevent
gevent-socketio
redis
numpy
matplotlib
request
tweepy
scikit-learn